+++ pkg/frontend/querymiddleware/roundtrip.go
@@ -267,6 +267,9 @@ func newQueryTripperware(
 	// This enables duration arithmetic https://github.com/prometheus/prometheus/pull/16249.
 	parser.ExperimentalDurationExpr = true
 
+	// This enables the anchored and smoothed selector modifiers
+	parser.EnableExtendedRangeSelectors = true
+
 	var c cache.Cache
 	if cfg.CacheResults || cfg.cardinalityBasedShardingEnabled() {
 		var err error
@@ -486,7 +489,7 @@ func newQueryMiddlewares(
 
 	// Does not apply to remote read as those are executed remotely and the enabling of PromQL experimental
 	// functions for those are not controlled here.
-	experimentalFunctionsMiddleware := newExperimentalFunctionsMiddleware(limits, log)
+	experimentalFunctionsMiddleware := newExperimentalFeaturesMiddleware(limits, log)
 	queryRangeMiddleware = append(
 		queryRangeMiddleware,
 		newInstrumentMiddleware("experimental_functions", metrics),