+++ pkg/streamingpromql/planning/core/vector_selector.go
@@ -22,7 +22,7 @@ type VectorSelector struct {
 }
 
 func (v *VectorSelector) Describe() string {
-	d := describeSelector(v.Matchers, v.Timestamp, v.Offset, nil, v.SkipHistogramBuckets)
+	d := describeSelector(v.Matchers, v.Timestamp, v.Offset, nil, v.SkipHistogramBuckets, false, v.Smoothed)
 
 	if v.ReturnSampleTimestamps {
 		return d + ", return sample timestamps"
@@ -70,7 +70,8 @@ func (v *VectorSelector) EquivalentToIgnoringHintsAndChildren(other planning.Nod
 		slices.EqualFunc(v.Matchers, otherVectorSelector.Matchers, matchersEqual) &&
 		((v.Timestamp == nil && otherVectorSelector.Timestamp == nil) || (v.Timestamp != nil && otherVectorSelector.Timestamp != nil && v.Timestamp.Equal(*otherVectorSelector.Timestamp))) &&
 		v.Offset == otherVectorSelector.Offset &&
-		v.ReturnSampleTimestamps == otherVectorSelector.ReturnSampleTimestamps
+		v.ReturnSampleTimestamps == otherVectorSelector.ReturnSampleTimestamps &&
+		v.Smoothed == otherVectorSelector.Smoothed
 }
 
 func (v *VectorSelector) MergeHints(other planning.Node) error {
@@ -99,6 +100,7 @@ func MaterializeVectorSelector(v *VectorSelector, _ *planning.Materializer, time
 		SkipHistogramBuckets:     v.SkipHistogramBuckets,
 		ExpressionPosition:       v.GetExpressionPosition().ToPrometheusType(),
 		MemoryConsumptionTracker: params.MemoryConsumptionTracker,
+		Smoothed:                 v.Smoothed,
 	}
 
 	return planning.NewSingleUseOperatorFactory(selectors.NewInstantVectorSelector(selector, params.MemoryConsumptionTracker, params.QueryStats, v.ReturnSampleTimestamps)), nil
@@ -109,8 +111,7 @@ func (v *VectorSelector) ResultType() (parser.ValueType, error) {
 }
 
 func (v *VectorSelector) QueriedTimeRange(queryTimeRange types.QueryTimeRange, lookbackDelta time.Duration) (planning.QueriedTimeRange, error) {
-	minT, maxT := selectors.ComputeQueriedTimeRange(queryTimeRange, TimestampFromTime(v.Timestamp), 0, v.Offset.Milliseconds(), lookbackDelta)
-
+	minT, maxT := selectors.ComputeQueriedTimeRange(queryTimeRange, TimestampFromTime(v.Timestamp), 0, v.Offset.Milliseconds(), lookbackDelta, false, v.Smoothed)
 	return planning.NewQueriedTimeRange(timestamp.Time(minT), timestamp.Time(maxT)), nil
 }
 
@@ -119,5 +120,8 @@ func (v *VectorSelector) ExpressionPosition() (posrange.PositionRange, error) {
 }
 
 func (v *VectorSelector) MinimumRequiredPlanVersion() planning.QueryPlanVersion {
+	if v.Smoothed {
+		return planning.QueryPlanV4
+	}
 	return planning.QueryPlanVersionZero
 }